requirements.txt
arrow==1.2.3
face_recognition==1.3.0
fasteners==0.18
fpdf==1.7.2
matplotlib==3.6.3
mysql_connector_repackaged==0.3.1
numpy==1.23.5
opencv_python_headless==4.7.0.68
openpyxl==3.0.10
pandas==1.5.2
pandas_profiling==3.6.2
Pillow
plotly==5.11.0
pymongo==4.3.3
python-dotenv==1.0.1
pytz==2022.7
SQLAlchemy==1.4.40
streamlit==1.12.0
streamlit_authenticator==0.2.1
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
vizpool
altair<5
mysql-connector
streamlit-aggrid==0.3.4.post3
pymysql==1.1.0
loguru
# XlsxWriter==3.2.0